# Make Uniform<f32/f64> and gen_range<f32/f64> honor limits strictly. Resolves #476