This adds a small amount of overhead to Uniform<f32/f64> and gen_range<f32/f64>. For Uniform<f32/f64> it adds one subtraction and one branch in the common case. For gen_range<f32/f64> it adds just one branch in the common case. I see no way around this while also being sure to stay in the requested range.

This does use f64::from_bits/to_bits/is_infinite/is_finite which might only be available through std? Though it looks like core::num also make them available [here](https://doc.rust-lang.org/stable/src/core/num/mod.rs.html#3591)? I can't tell given that the functions are marked as stable, but the trait marked unstable.

I'm not sure what are ok build requirements for no_std builds? And is there any way to run `cargo test`, or even just `cargo build` in no_std mode?