@pitdicker 2018-05-25
I understand your motivation, but I am not sure I agree with the method.

It is not impossible to make a half-open range exact, i.e. to guarantee `low` can occur and `high` never does. It could be done as a check when constructing the range that `low + scale * (1 - EPSILON)` (or some other variant, didn't look closely at the code yet) is less than `high`, and to otherwise adjust `scale`.

That could be faster for many samples, but slower for few or only one.

Can you measure one thing: what do the range benchmarks do with the extra loop?