@sicking 2018-05-25
Given the low performance cost here I think it's worth providing the guarantee that the generated values are within the requested range.

I've implemented @pitdicker suggestion to adjust `scale` in `UniformFloat::new`/ `UniformFloat::new_inclusive`, which means that the performance cost when sampling from a range is just one floating subtraction, and the cost when using `Rng.gen_range` is just one branch.

I'll pull numbers to see if this is visible in our benchmarks.