@sicking 2018-05-26
So I ran some performance numbers. At first `distr_uniform_f32` showed a 6.0% slowdown, and `distr_uniform_f64` showed a 3.8%. Which to me seems quite acceptable.

Then I added `gen_range_f32` and `gen_range_f64`, using the exact same code as in `gen_range_i8`. Here `gen_range_f32` showed almost a 2x **speedup** with this PR, while `gen_range_f64` showed almost a 2x **slowdown** with the PR. No idea what's going on there!

Also confused why some builds seem unable to find the `black_box` function, even though it's used just fine elsewhere in the same file. Will look into that after the weekend.