@pitdicker 2018-05-28
Sorry, I've been a bit slow this weekend. Will have a better look tomorrow.

We had trouble with confusing benchmarks in combination with `XorShiftRng` before. The assembly is even identical, still it can perform very differently... :slightly_frowning_face:

The best thing to do than is to try again with `StdRng`. Benchmarks before:
```
test gen_range_f32               ... bench:       3,388 ns/iter (+/- 54) = 1180 MB/s
test gen_range_f64               ... bench:       5,000 ns/iter (+/- 317) = 1600 MB/s
```
After:
```
test gen_range_f32               ... bench:       3,800 ns/iter (+/- 5) = 1052 MB/s
test gen_range_f64               ... bench:       5,412 ns/iter (+/- 4) = 1478 MB/s
```

So the difference in both benchmarks is 0.412 ns. `distr_uniform_f32` was 1.8 ns/iter (with Xorshift) on my machine, so about 22% overhead. And for the `f64` version about 16%.

> Also confused why some builds seem unable to find the `black_box` function, even though it's used just fine elsewhere in the same file. Will look into that after the weekend.

I find it strange also, but would not worry about it. If you rebase on master you'll see all the `black_box`-es are unnecessary now.