@sicking 2018-06-04
Yeah, using `StdRng` I get similar numbers:

Before:
```
test gen_range_f32               ... bench:       2,870 ns/iter (+/- 62) = 1393 MB/s
test gen_range_f64               ... bench:       4,213 ns/iter (+/- 82) = 1898 MB/s
test distr_uniform_f32           ... bench:       1,497 ns/iter (+/- 44) = 2672 MB/s
test distr_uniform_f64           ... bench:       2,099 ns/iter (+/- 56) = 3811 MB/s
```

After:
```
test gen_range_f32               ... bench:       3,781 ns/iter (+/- 198) = 1057 MB/s
test gen_range_f64               ... bench:       4,940 ns/iter (+/- 415) = 1619 MB/s
test distr_uniform_f32           ... bench:       1,614 ns/iter (+/- 31) = 2478 MB/s
test distr_uniform_f64           ... bench:       2,188 ns/iter (+/- 30) = 3656 MB/s
```

So a 32% slowdown for `gen_range_f32`, and a 17% for `gen_range_f64`.

But for the more performance-optimized API of creating a `Uniform<f32/f64>` the slowdown is 7.8% for `distr_uniform_f32` and 4.2% for `distr_uniform_f64`.

Definitely more slowdown than desired for `gen_range`, presumably due to the added branch getting mispredicted, but I think the added correctness is worth it.