@sicking 2018-06-04
As a test, I tried adding branch-predictor hints to the `gen_range` code path, and that significantly brought down the overhead. Now I get:

```
test gen_range_f32               ... bench:       3,533 ns/iter (+/- 70) = 1132 MB/s
test gen_range_f64               ... bench:       4,447 ns/iter (+/- 74) = 1798 MB/s
```

Which is a 23% slowdown for `gen_range_f32` and 5.5% slowdown for `gen_range_f64`. Unfortunately the branch-predictor hinting API isn't at all stable yet, so not something we can use right now.