@sicking 2018-06-08
Here's an alternative approach for implementing `UniformFloat::new` which would let us keep `Uniform::sample` as-is and avoid the perf-hit.

However it seems hangs for really large values of `high`. I think the fact that both `offset` and `scale` depend on both `high` and `low`, in some cases the algorithm below ends up adjusting the wrong edge. Most likely related to when `scale` and/or `offset` end up being infinite and/or NaN.

I'm sure it can be fixed, but ultimately I don't think the complexity and precision loss is worth it. And the algorithm here has effectively unbounded performance cost, which is also a concern.

So I think we should stick with the approach in the PR.

```rust
fn new(low: Self::X, high: Self::X) -> Self {
    assert!(low < high, "Uniform::new called with `low >= high`");
    assert!(low.is_finite() && high.is_finite(), "Uniform::new called with non-finite boundaries");

    fn adjust_down(val: $ty) -> $ty {
        if val > 0.0 {
            <$ty>::from_bits(val.to_bits() - 1)
        } else if val < 0.0 {
            <$ty>::from_bits(val.to_bits() + 1)
        } else {
            -<$ty>::from_bits(1)
        }
    }

    let max_rand = (::core::$uty::MAX >> $bits_to_discard)
                   .into_float_with_exponent(0);
    let min_rand = (0 as $uty >> $bits_to_discard)
                   .into_float_with_exponent(0);

    let mut low_adjusted = low;
    let mut high_adjusted = high;
    loop {
        let scale = high_adjusted - low_adjusted;
        let offset = low_adjusted - scale;
        if !(scale * max_rand + offset < high) {
            high_adjusted = adjust_down(high_adjusted);
        } else if !(scale * min_rand + offset >= low) {
            low_adjusted = -adjust_down(-low_adjusted);
        } else {
            return UniformFloat { scale, offset };
        }
    }
}
```