@sicking 2018-07-02
I've rebased this on top of the newly landed SIMD code, which was definitely non-trivial. The good news is that it's enforcing the limits for both primitives and SIMD types.

I also found a way to make `sample_single` faster, so it should be less of a perf-hit compared to what we're currently doing.

I'd really like to get this landed though. Especially since the weighted sampling has landed and currently risks panicing randomly if you use f32/f64 weights.