@dhardy 2018-07-12
Okay, I think it's time I stopped opposing this based on performance overhead; it's barely above the noise threshold (notice how unrelated numbers change here):
```
# before
test gen_range_f32               ... bench:       4,259 ns/iter (+/- 138) = 939 MB/s
test gen_range_f64               ... bench:       4,034 ns/iter (+/- 148) = 1983 MB/s
test gen_range_i16               ... bench:       5,009 ns/iter (+/- 352) = 399 MB/s
test gen_range_i32               ... bench:       2,892 ns/iter (+/- 81) = 1383 MB/s
test gen_range_i64               ... bench:       6,663 ns/iter (+/- 177) = 1200 MB/s
test gen_range_i8                ... bench:       4,802 ns/iter (+/- 87) = 208 MB/s
# after
test gen_range_f32               ... bench:       4,806 ns/iter (+/- 302) = 832 MB/s
test gen_range_f64               ... bench:       4,523 ns/iter (+/- 197) = 1768 MB/s
test gen_range_i16               ... bench:       4,794 ns/iter (+/- 100) = 417 MB/s
test gen_range_i32               ... bench:       2,997 ns/iter (+/- 458) = 1334 MB/s
test gen_range_i64               ... bench:       6,632 ns/iter (+/- 223) = 1206 MB/s
test gen_range_i8                ... bench:       5,033 ns/iter (+/- 62) = 198 MB/s
```

Is this ready?

Please rebase, and please separate the new benchmarks into a separate commit (so it's easier to do before/after comparisons).